Create README.md
This program was created in Visual Studio 2019 in conjunction to a video tutorial. I'm new to coding in my internship. It has a lot of errors in it and not function like the video shows at this resource: https://www.youtube.com/watch?v=kbAksctUEzc Any help or advice on some of the errors would be appreciated.